requirements.txt
requests
psutil
pymysql
PyYAML
cryptography
